Add kinematic rigidbody and collider to portal
The portal doesn't move, but OnTriggerEnter2D/OnTriggerExit2D require at least one associated collider to have a rigidbody. We don't want to attach a rigidbody directly to the range prefab (since once of its parents might have a rigidbody, as would be the case with mobs), so we have to attach one to the structure.