Base whether Connectors can connect on object center
Instead of checking whether any part of the "hitbox collider" (the main collider attached to the top-level object) is within range of another Connector, check only the object's center point.

This makes connections a bit easier to reason about because we do not have to support arbitrary colliders. This should also make it easier to unify the connection checking code in PlayerController and Connector.