Fix AIMovement reporting no target too soon
Previously, HasTarget() returned false as soon as the path ran out of nodes. The issue was that the AI could still be in the process of moving toward the last node.

This was not noticeable for walking enemies because they tended to stop before the last node was popped from the path stack. Flying enemies, however, only have one node in their path to start with, resulting in the path being cleared immediately.